Big clean: arch train and test is working
