+++ Evals/ArchComparison_1.csv
@@ -1,0 +1,6 @@
+#,Name,%Complete,AvgTime,Std,NoObs
+0,Navforest_1,10.0,783.0,0.0,940.0
+1,Follow the Forest Gap,100.0,441.4,42.673645262620816,373.0
+2,Oracle Path Follower,100.0,384.7,18.30874108178932,372.0
+3,ModForest_1,40.0,681.0,95.55626614722867,667.0
+4,SapForest_dist_1,100.0,413.6,15.311433636338565,372.0
