--- TestingScripts/TestLPD.py
@@ -1,229 +1,0 @@
-
-
-from LearningLocalPlanning.NavAgents.AgentNav import NavTestVehicle
-from LearningLocalPlanning.NavAgents.AgentMod import ModVehicleTest 
-from LearningLocalPlanning.NavAgents.Oracle import Oracle
-from LearningLocalPlanning.NavAgents.FollowTheGap import ForestFGM
-from LearningLocalPlanning.NavAgents.SerialAgentPlanner import SerialVehicleTest
-
-
-from LearningLocalPlanning.Simulator.ForestSim import ForestSim
-
-import numpy as np 
-import csv, yaml 
-from argparse import Namespace
-
-map_name = "forest2"
-train_n = 1
-nav_name = f"Navforest_{train_n}"
-mod_name = f"ModForest_{train_n}"
-sap_safe_name = f"SapForest_safe_{train_n}"
-sap_cth_name = f"SapForest_cth_{train_n}"
-sap_dist_name = f"SapForest_dist_{train_n}"
-
-eval_name = f"CompareTestLPD_{train_n}"
-
-n_test = 1000
-
-
-def load_conf(path, fname):
-    full_path = path + 'config/' + fname + '.yaml'
-    with open(full_path) as file:
-        conf_dict = yaml.load(file, Loader=yaml.FullLoader)
-
-    conf = Namespace(**conf_dict)
-
-    return conf
-
-
-
-
-"""Testing Function"""
-class TestData:
-    def __init__(self) -> None:
-        self.endings = None
-        self.crashes = None
-        self.completes = None
-        self.lap_times = None
-        self.lap_times_no_obs = None
-
-        self.names = []
-        self.lap_histories = None
-
-        self.N = None
-
-    def init_arrays(self, N, laps):
-        self.completes = np.zeros((N))
-        self.crashes = np.zeros((N))
-        self.lap_times = np.zeros((laps, N))
-        self.lap_times_no_obs = np.zeros((N))
-        self.endings = np.zeros((laps, N)) #store env reward
-        self.lap_times = [[] for i in range(N)]
-        self.N = N
- 
-    def save_txt_results(self):
-        test_name = 'Evals/' + self.eval_name + '.txt'
-        with open(test_name, 'w') as file_obj:
-            file_obj.write(f"\nTesting Complete \n")
-            file_obj.write(f"Map name:  \n")
-            file_obj.write(f"-----------------------------------------------------\n")
-            file_obj.write(f"-----------------------------------------------------\n")
-            for i in range(self.N):
-                file_obj.write(f"Vehicle: {self.vehicle_list[i].name}\n")
-                file_obj.write(f"Crashes: {self.crashes[i]} --> Completes {self.completes[i]}\n")
-                percent = (self.completes[i] / (self.completes[i] + self.crashes[i]) * 100)
-                file_obj.write(f"% Finished = {percent:.2f}\n")
-                file_obj.write(f"Avg lap times: {np.mean(self.lap_times[i])}\n")
-                file_obj.write(f"No Obs Time: {self.lap_times_no_obs[i]}\n")
-
-                file_obj.write(f"-----------------------------------------------------\n")
-
-    def print_results(self):
-        print(f"\nTesting Complete ")
-        print(f"-----------------------------------------------------")
-        print(f"-----------------------------------------------------")
-        for i in range(self.N):
-            if len(self.lap_times[i]) == 0:
-                self.lap_times[i].append(0)
-            print(f"Vehicle: {self.vehicle_list[i].name}")
-            print(f"Crashes: {self.crashes[i]} --> Completes {self.completes[i]}")
-            percent = (self.completes[i] / (self.completes[i] + self.crashes[i]) * 100)
-            print(f"% Finished = {percent:.2f}")
-            print(f"Avg lap times: {np.mean(self.lap_times[i])}")
-            print(f"No Obs Time: {self.lap_times_no_obs[i]}")
-            print(f"-----------------------------------------------------")
-        
-    def save_csv_results(self):
-        test_name = 'Evals/'  + self.eval_name + '.csv'
-
-        data = [["#", "Name", "%Complete", "AvgTime", "Std", "NoObs"]]
-        for i in range(self.N):
-            v_data = [i]
-            v_data.append(self.vehicle_list[i].name)
-            v_data.append((self.completes[i] / (self.completes[i] + self.crashes[i]) * 100))
-            v_data.append(np.mean(self.lap_times[i]))
-            v_data.append(np.std(self.lap_times[i]))
-            v_data.append(self.lap_times_no_obs[i])
-            data.append(v_data)
-
-        with open(test_name, 'w') as csvfile:
-            csvwriter = csv.writer(csvfile)
-            csvwriter.writerows(data)
-
-
-
-class TestVehicles(TestData):
-    def __init__(self, config, eval_name, env_kwarg='forest') -> None:
-        self.config = config
-        self.eval_name = eval_name
-        self.vehicle_list = []
-        self.N = None
-        self.env_kwarg = env_kwarg
-
-        TestData.__init__(self)
-
-    def add_vehicle(self, vehicle):
-        self.vehicle_list.append(vehicle)
-
-    def run_eval(self, env, laps=100, show=False, wait=False):
-        N = self.N = len(self.vehicle_list)
-        self.init_arrays(N, laps)
-
-        # No obstacles
-        for j in range(N):
-            vehicle = self.vehicle_list[j]
-
-            r, steps = self.run_lap(vehicle, env, show, False, wait)
-            self.lap_times_no_obs[j] = env.steps
-
-            print(f"#NoObs: Lap time for ({vehicle.name}): {env.steps} --> Reward: {r}")
-
-        for i in range(laps):
-            env.env_map.add_obstacles()
-            for j in range(N):
-                vehicle = self.vehicle_list[j]
-
-                r, steps = self.run_lap(vehicle, env, show, False, wait)
-
-                print(f"#{i}: Lap time for ({vehicle.name}): {env.steps} --> Reward: {r}")
-                self.endings[i, j] = r
-                if r == -1 or r == 0:
-                    self.crashes[j] += 1
-                else:
-                    self.completes[j] += 1
-                    self.lap_times[j].append(steps)
-
-        self.print_results()
-        self.save_txt_results()
-        self.save_csv_results()
-
-    def run_lap(self, vehicle, env, show, add_obs, wait):
-        env.scan_sim.reset_n_beams(vehicle.n_beams)
-        state = env.reset(add_obs)
-
-        try:
-            vehicle.plan_forest(env.env_map)
-        except AttributeError as e:
-            pass
-
-        done = False
-        while not done:
-            a = vehicle.plan_act(state)
-            s_p, r, done, _ = env.step_plan(a)
-            state = s_p
-            # env.render(False)
-
-        if show:
-            # vehicle.show_vehicle_history()
-            # env.history.show_history()
-            if wait:
-                env.render(wait=True, name=vehicle.name)
-            else:
-                env.render(wait=False, name=vehicle.name)
-
-        return r, env.steps
-
-
-
-
-def comparison_test():
-    sim_conf = load_conf("", "std_config")
-    env = ForestSim(map_name, sim_conf)
-    test = TestVehicles(sim_conf, eval_name)
-
-    vehicle = NavTestVehicle(nav_name, sim_conf)
-    test.add_vehicle(vehicle)
-
-    vehicle = ForestFGM()
-    test.add_vehicle(vehicle)
-
-    vehicle = Oracle(sim_conf)
-    test.add_vehicle(vehicle)
-
-    vehicle = ModVehicleTest(mod_name, map_name, sim_conf)
-    test.add_vehicle(vehicle)
-
-    vehicle = SerialVehicleTest(sap_cth_name, map_name, sim_conf)
-    test.add_vehicle(vehicle)
-    vehicle = SerialVehicleTest("TuneCTH_0004_0002", map_name, sim_conf)
-    test.add_vehicle(vehicle)
-    vehicle = SerialVehicleTest("TuneCTH_0004_001", map_name, sim_conf)
-    test.add_vehicle(vehicle)
-
-    vehicle = SerialVehicleTest(sap_safe_name, map_name, sim_conf)
-    test.add_vehicle(vehicle)
-
-    vehicle = SerialVehicleTest(sap_dist_name, map_name, sim_conf)
-    test.add_vehicle(vehicle)
-
-    # test.run_eval(env, 1, True)
-    test.run_eval(env, n_test, False, wait=False)
-
-
-
-if __name__ == "__main__":
-    
-
-    comparison_test()
-
-
